Numerous changes to variable types representing the length of a memory buffer/cstring; always use size_t for this purpose.

+++ doomsday/engine/api/dd_api.h
@@ -82,7 +82,7 @@ typedef struct {
     long int      (*NetPlayerEvent) (int playernum, int type, void *data);
     int           (*NetWorldEvent) (int type, int parm, void *data);
     void          (*HandlePacket) (int fromplayer, int type, void *data,
-                                   int length);
+                                   size_t length);
 
     // Tickers.
     void          (*Ticker) (timespan_t ticLength);
