+++ doomsday/engine/portable/include/m_misc.h
@@ -79,9 +79,9 @@ boolean         M_IsComment(char *text);
 boolean         M_IsStringValidInt(const char *str);
 boolean         M_IsStringValidByte(const char *str);
 boolean         M_IsStringValidFloat(const char *str);
-char           *M_LimitedStrCat(const char *str, unsigned int maxWidth,
+char           *M_LimitedStrCat(const char *str, size_t maxWidth,
                                 char separator, char *buf,
-                                unsigned int bufLength);
+                                size_t bufLength);
 char           *M_StrCatQuoted(char *dest, char *src);
 
 // Random numbers.
